budb: Add printf format checks to logging fns
Add format checking to the Log, LogError and LogDebug functions.

Fix the resulting type errors. One of these in particular is a
signficant bug, where we would try to print a hash table pointer
as a string.

Reviewed-on: http://gerrit.openafs.org/2445
Tested-by: Derrick Brashear <shadow@dementia.org>

+++ src/budb/db_alloc.c
@@ -206,8 +206,8 @@ AllocStructure(struct ubik_trans *ut, char type, dbadr related, dbadr *saP, void
     *(afs_int32 *) s = 1;	/* make sure structure is not free */
     *saP = a + ((char *)bs - (char *)&b);
 
-    LogDebug(3, "allocated at %d, block at %d, offset %d\n", *saP, a,
-	     ((char *)bs - (char *)&b));
+    LogDebug(3, "allocated at %d, block at %d, offset %ld\n", *saP, a,
+	     (long int)((char *)bs - (char *)&b));
     /* caller must write back at least first word of structure */
     return 0;
 }
